migrate prog to main
